Resolve various issues to ensure buttons shows for Jira
Fixes #33

It was found a few months ago that the Clockify button stopped appearing
in Jira. Some time was taken to try to resolve any issues in the 4 main
views to ensure the button would show in the correct place, with the
correct issue number identified.

Work in this PR was tested in both Firefox and Chrome by using
`debugger;` and inspecting what was found in all 4 issue views.

The main issue in Issue #33 was that `project` was not getting found
before child nodes were accessed. I'm not fully clear on where it's
trying to find a project, _but_ adding some checks before accessing
child nodes resolved the error on my side.

After that was resolved, selectors and styles were updated to ensure the
Clockify button was added in the correct place without breaking the UI.
Additionally, some updates were done to ensure the correct Jira issue
number is found when an issue is by itself or within an epic.