Add freshservice integration (#147)
* Add Freshservice integration

Freshservice is a help-desk product in the FreshWorks family, alongside
Freshdesk and Freshrelease.  This commit adds the start/stop timer
button to the ticket detail page.

On Chrome, Freshservice's CSS overrides the left margin of span
elements, resulting in incorrect padding between the button icon and
text.  Adding the !important flag to button span's CSS corrects this.

* Drive-by: Fix mismatched webpack-dev-middleware dependency

Running `npm install` was resulting in the following error:

    Could not resolve dependency:
    peer webpack@"^1.0.0 || ^2.0.0 || ^3.0.0" from webpack-dev-middleware@1.12.2
    node_modules/webpack-dev-middleware
      dev webpack-dev-middleware@"^1.10.2" from the root project

Upgrading webpack-dev-middleware to the latest version seems to correct
the problem.